Stop the static_cast<> & size_t madness!  (at least in buffered_resource_loader.cc)
Review URL: http://codereview.chromium.org/10052031

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@131936 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: f6ee0f67d4652e94b614ffcc8ee532ab99f2847c

+++ webkit/media/buffered_resource_loader.h
@@ -286,7 +286,7 @@ class BufferedResourceLoader : public WebKit::WebURLLoaderClient {
   // read has completed or failed.
   ReadCB read_cb_;
   int64 read_position_;
-  size_t read_size_;
+  int read_size_;
   uint8* read_buffer_;
 
   // Offsets of the requested first byte and last byte in |buffer_|. They are
